+++ src/components/chip/chip.styles.ts
@@ -26,31 +26,31 @@ export default css`
 
   .chip--neutral {
     border-color: var(--cps-color-system-neutral-transparent);
-    background-color: var(--cps-color-system-neutral-subtle);
+    background-color: var(--cps-color-state-neutral-subtle);
     color: var(--cps-palette-neutral-800);
   }
 
   .chip--informative {
     border-color: var(--cps-color-stroke-primary);
-    background-color: var(--cps-color-system-informative);
+    background-color: var(--cps-color-state-informative);
     color: var(--cps-palette-neutral-0);
   }
 
   .chip--warning {
     border-color: var(--cps-color-stroke-primary);
-    background-color: var(--cps-color-system-warning);
+    background-color: var(--cps-color-state-warning);
     color: var(--cps-palette-neutral-0);
   }
 
   .chip--critical {
     border-color: var(--cps-color-stroke-primary);
-    background-color: var(--cps-color-system-critical);
+    background-color: var(--cps-color-state-critical);
     color: var(--cps-palette-neutral-0);
   }
 
   .chip--success {
     border-color: var(--cps-color-stroke-primary);
-    background-color: var(--cps-color-system-success);
+    background-color: var(--cps-color-state-success);
     color: var(--cps-palette-neutral-0);
   }